Use `u32` for `Index`

+++ fj/src/graphics/renderer.rs
@@ -290,7 +290,7 @@ impl Renderer {
             render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
             render_pass.set_index_buffer(
                 self.index_buffer.slice(..),
-                wgpu::IndexFormat::Uint16,
+                wgpu::IndexFormat::Uint32,
             );
             render_pass.draw_indexed(0..self.num_indices, 0, 0..1);
         }
